+++ bee-ledger/bee-ledger-types/CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 1.0.0-beta.2 - 2022-07-20
+
+### Changed
+
+- Bump `bee-block` dependency;
+
 ## 1.0.0-beta.1 - 2022-07-17
 
 First alpha release.